Remove option, revert DB change, move query to TimetableDayDateGateway
